refactor: improve exported types
