+++ lib/endguard.dart
@@ -1,7 +1,6 @@
 library cryptography;
 
 export 'src/connection/connection.dart';
-export 'src/handshake/exception.dart';
+export 'src/crypto/exception.dart';
 export 'src/handshake/message.dart';
 export 'src/parallelism/exception.dart';
-export 'src/protos/protocol.pb.dart';
