+++ lib/src/parallelism/exception.dart
@@ -1,16 +1,8 @@
+import 'package:endguard/src/parallelism/operations.dart';
 import 'package:endguard/src/protos/protocol.pb.dart';
 
-export 'package:endguard/src/protos/protocol.pb.dart' show State;
-
-enum Operation {
-  ApplyConnectionConfirmation,
-  ApplyConnectionOffer,
-  CreateConnectionOffer,
-  DecryptMessage,
-  EncryptMessage,
-  ExportState,
-  UpdateConnectionSettings,
-}
+export 'package:endguard/src/parallelism/operations.dart' show Operation;
+export 'package:endguard/src/protos/protocol.pb.dart' show ConnectionState_State;
 
 String _operationDescription(Operation op) {
   switch (op) {
