+++ lib/src/parallelism/operations.dart
@@ -1,0 +1,9 @@
+enum Operation {
+  ApplyConnectionConfirmation,
+  ApplyConnectionOffer,
+  CreateConnectionOffer,
+  DecryptMessage,
+  EncryptMessage,
+  ExportState,
+  UpdateConnectionSettings,
+}
